requirements.txt
# Core Framework
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1

# Database - SQLite (já incluso no Python) ou PostgreSQL
SQLAlchemy==2.0.21
alembic==1.12.1
python-dotenv==1.0.0

# PostgreSQL (opcional - se for usar no futuro)
# psycopg2-binary==2.9.9

# Timezone and Date
pytz==2023.3

# Device Detection
user-agents==2.2.0

# HTTP and Security
requests==2.31.0
bleach==6.1.0
email-validator==2.1.0
werkzeug==2.3.7

# Development tools
watchdog==3.0.0